Update: Get updated frame count every second. Display progress bar
Use regex to get the updated number of frames every second. Once the
frames are fetched, the thread sleeps for one second to avoid wasting
CPU power.

Add method to print the percentage of progress, a progress bar to
indicate the same and a counter to display approximate remaining time to
generate the FLAC file.