Update: Connect all functions. Get the script working.
Modify the `generate_flac_file` method to return a tuple instead of a
boolean. This tuple will contain the boolean as well as a string. The
string will contain full path of the flac file generated.

Once user enters an input, generate a flac file for all the files
present inside the root directory.