Update: Handle edge cases where total frame count can't be fetched
Add exception handling to handle the edge cases where ffmpeg output does
not contain the count for the total number of frames present in the
file.

In case frame count can't be read, script won't display progress bar,
the percentage of progress or the time remaining. Will simply display
count for the current number of frames in such a scenario.