Minor: Display real-time progress smoothly.
Handle displaying the progress bar with the current progress percentage.
The length of the string printed while displaying the progress will be
fixed.

The progress percentage will be padded with zeroes on the right and
spaces on the left in order to make it fit the required length.

Once a flac is created successfully the progress bar won't be
overwritten with the success message, instead the progress bar will be
left untouched with the bar filled and time remaining as zero and the
success message will be printed on a new line.