Patch: Patch a few bugs, minor refactoring. Add support for linux
This commit is mainly to patch a few minor bugs that I planned to
patch a long time ago, apart from this, it also includes a bit
of refactoring - including improvemenets in internal code comments
explaining stuff in a simpler manner.

Another big change being introduced in this commit is the support
for linux, back when I initially created this project, I made it to
run on my Windows system, this commit adds support for Linux-based
system(s) too.